pyproject.toml
[tool.poetry]
name = "event_planner"
version = "2023.12.21"
description = ""
authors = [
  "Xavier Francisco <xavierfrancisco@flawlessworkflow.com>",
]

[tool.poetry.dependencies]
python = "^3.11.0"
sentry-sdk = "^1.40.0"
psycopg2-binary = "^2.9.9"
Django = "^5.0.1"
djangorestframework = "^3.14.0"
django-filter = "^23.5"
django-cors-headers = "^4.3.1"
dj-database-url = "^2.1.0"
django-split-settings = "^1.2.0"
drf-spectacular = "^0.26.4"
gunicorn = "^21.2.0"
django-stubs-ext = "^4.2.7"
dj-rest-auth = {extras = ["with-social"], version = "==4.0.1"}
djangorestframework-simplejwt = "^5.3.1"
rest-framework-mixins = "^1.0.0"

[tool.poetry.group.dev.dependencies]
tblib = "^3.0.0"                        # Added for tracebacks when running manage.py test --parallel
ruff = "^0.2.0"
ipdb = "^0.13.13"
ipython = "^8.21.0"
bumpver = "^2023.1129"
black = "^24.1.1"
pre-commit = "^3.6.0"
mypy = "^1.8.0"
typing-extensions = "^4.9.0"
types-dj-database-url = "^1.3.0.4"
types-toml = "^0.10.8.7"
django-stubs = "^4.2.7"
django-filter-stubs = "^0.1.3"
djangorestframework-stubs = "^3.14.5"
pylint = "^3.0.3"
pylint-django = "^2.5.5"
absolufy-imports = "^0.3.1"
django-extensions = "^3.2.3"
Werkzeug = "^3.0.1"
factory-boy = "^3.3.0"
coverage = "^7.4.1"
whitenoise = "^6.6.0"
tqdm = "^4.66.1"
line-profiler = "^4.1.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.flawless]
poetry-version = "1.7.0"
